Refactor frontend scripts into modular files
