+++ frontend/public/scripts/config.js
@@ -1,0 +1,16 @@
+export const DEFAULT_BACKEND_URL = "http://localhost:3000";
+
+export const TALKING_STREAK_FIELD = "talking-streak";
+export const USERNAME_STORAGE_KEY = "poke-pet.username";
+export const MS_PER_DAY = 24 * 60 * 60 * 1000;
+
+export const STAGE_DETAILS = {
+  1: { species: "Pichu", image: "images/pichu.png" },
+  2: { species: "Pikachu", image: "images/pikachu.png" },
+  3: { species: "Raichu", image: "images/raichu.png" },
+};
+
+export const DEFAULT_STAGE_DETAIL = {
+  species: "Companion",
+  image: "images/pikachu.png",
+};
