+++ frontend/public/scripts/state.js
@@ -1,0 +1,7 @@
+export const displayMessages = [];
+export const conversationHistory = [];
+
+export function resetState() {
+  displayMessages.splice(0);
+  conversationHistory.splice(0);
+}
