+++ frontend/public/scripts/ui/profile.js
@@ -1,0 +1,131 @@
+import { createElement } from "../dom.js";
+import { getStageDetail, selectActivePet } from "../pets.js";
+import { clampFriendship, getTalkingStreakValue, sanitizeIdentifier } from "../utils.js";
+
+function createInfoRow(label, value) {
+  return createElement("div", {
+    className: "info-row",
+    children: [
+      createElement("span", { className: "info-label", textContent: label }),
+      createElement("span", { className: "info-value", textContent: value }),
+    ],
+  });
+}
+
+function buildFriendshipSection(friendshipValue) {
+  const clampedValue = clampFriendship(friendshipValue);
+
+  const wrapper = createElement("div", { className: "friendship-wrapper" });
+  const header = createElement("div", {
+    className: "info-row",
+    children: [
+      createElement("span", { className: "info-label", textContent: "Friendship" }),
+      createElement("span", {
+        className: "friendship-value",
+        textContent: `${clampedValue}/100`,
+      }),
+    ],
+  });
+  const bar = createElement("div", { className: "friendship-bar" });
+  const fill = createElement("div", { className: "friendship-bar-fill" });
+  fill.style.width = `${clampedValue}%`;
+  bar.appendChild(fill);
+  wrapper.appendChild(header);
+  wrapper.appendChild(bar);
+
+  return wrapper;
+}
+
+function resolveLastChattedDisplay(lastChattedRaw) {
+  if (typeof lastChattedRaw === "string" && lastChattedRaw.trim()) {
+    const trimmed = lastChattedRaw.trim();
+    const parsed = new Date(trimmed);
+    return Number.isNaN(parsed.getTime()) ? trimmed : parsed.toISOString().split("T")[0];
+  }
+
+  if (lastChattedRaw instanceof Date) {
+    return lastChattedRaw.toISOString().split("T")[0];
+  }
+
+  if (lastChattedRaw !== undefined && lastChattedRaw !== null) {
+    const parsed = new Date(lastChattedRaw);
+
+    if (!Number.isNaN(parsed.getTime())) {
+      return parsed.toISOString().split("T")[0];
+    }
+  }
+
+  return "—";
+}
+
+export function buildProfileColumn({ user, pet }) {
+  const column = createElement("div", { className: "profile-column" });
+  const activePet = pet ?? selectActivePet(user);
+  const stageDetail = getStageDetail(activePet?.stage);
+
+  const petName =
+    activePet && typeof activePet.name === "string" && activePet.name.trim()
+      ? activePet.name.trim()
+      : "Your Companion";
+
+  const avatar = createElement("img", {
+    className: "pet-avatar",
+    attributes: {
+      src: stageDetail.image,
+      alt: `${stageDetail.species} avatar`,
+    },
+  });
+
+  const petCardChildren = [
+    avatar,
+    createElement("div", { className: "pet-name", textContent: petName }),
+  ];
+
+  if (!activePet) {
+    petCardChildren.push(
+      createElement("div", {
+        className: "info-value",
+        textContent: "Add a pet to begin chatting.",
+      }),
+    );
+  }
+
+  column.appendChild(
+    createElement("div", {
+      className: "card pet-card",
+      children: petCardChildren,
+    }),
+  );
+
+  const speciesFromPet =
+    activePet && typeof activePet.species === "string" && activePet.species.trim()
+      ? activePet.species.trim()
+      : stageDetail.species;
+
+  const ownerId = sanitizeIdentifier(user.id, "Player");
+
+  const lastChattedDisplay = resolveLastChattedDisplay(activePet?.lastChatted);
+
+  const infoRows = [
+    createInfoRow("Species", speciesFromPet),
+    createInfoRow("Owner", ownerId),
+    createInfoRow("Last-chatted", lastChattedDisplay),
+  ];
+
+  const talkingStreakDisplay = `${getTalkingStreakValue(activePet)} days`;
+  infoRows.push(createInfoRow("Talking Streak", talkingStreakDisplay));
+
+  const infoCardChildren = [
+    createElement("div", { className: "info-grid", children: infoRows }),
+    buildFriendshipSection(activePet?.friendship ?? 0),
+  ];
+
+  column.appendChild(
+    createElement("div", {
+      className: "card info-card",
+      children: infoCardChildren,
+    }),
+  );
+
+  return column;
+}
